requirements.txt
Flask==1.1.2
http-parser==0.9.0
itsdangerous==1.1.0
prettytable==2.1.0
psycopg2-binary==2.8.6
http_parser
